ADD/FIX: metod unpairVehicle() and other things
Hemos cambiado de double a float en varias variables, ya que esperaba un float y no un double.
 Hemos añadido el setId del vehiculo en "SqanQR".
Hemos cambiado de string a VehicleID en el método getId()

+++ MicromobilidadCompartida/src/main/java/data/GeographicPoint.java
@@ -34,22 +34,22 @@ final public class GeographicPoint {
                 "longitude='" + longitude + "\'}";
     }
 
-    public double calculateDistance(GeographicPoint other) {
-        double R = 6371.0; // Earth's radius
+    public float calculateDistance(GeographicPoint other) {
+        float R = 6371; // Earth's radius
 
-        double lat1 = Math.toRadians(this.latitude);
-        double lon1 = Math.toRadians(this.longitude);
-        double lat2 = Math.toRadians(other.latitude);
-        double lon2 = Math.toRadians(other.longitude);
+        float lat1 = (float) Math.toRadians(this.latitude);
+        float lon1 = (float) Math.toRadians(this.longitude);
+        float lat2 = (float) Math.toRadians(other.latitude);
+        float lon2 = (float) Math.toRadians(other.longitude);
 
-        double latDistance = lat2 - lat1;
-        double lonDistance = lon2 - lon1;
+        float latDistance = lat2 - lat1;
+        float lonDistance = lon2 - lon1;
 
         // Haversine formula
-        double a = Math.sin(latDistance / 2) * Math.sin(latDistance / 2) + Math.cos(lat1) * Math.cos(lat2) *
-                   Math.sin(lonDistance / 2) * Math.sin(lonDistance / 2);
+        float a = (float) (Math.sin(latDistance / 2) * Math.sin(latDistance / 2) + Math.cos(lat1) * Math.cos(lat2) *
+                Math.sin(lonDistance / 2) * Math.sin(lonDistance / 2));
 
-        double c = 2 * Math.atan2(Math.sqrt(a), Math.sqrt(1 - a));
+        float c = (float) (2 * Math.atan2(Math.sqrt(a), Math.sqrt(1 - a)));
 
         // La distancia en kilómetros
         return R * c;
